INTLY-1362 Update walkthrough loading screen to re-use startup loading screen
